image-search-app/requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imgbeddings==0.1.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==33.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
mpmath==1.3.0
narwhals==1.47.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
open_clip_torch==2.32.0
openai==1.97.0
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentencepiece==0.2.0
sentry-sdk==2.33.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
stack-data==0.6.3
starlette==0.47.1
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.17
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
zipp==3.23.0